*裁图 看到 手势添加 *个人仓库代码优化

+++ MyDevelopmentTest/MyDevelopmentTest/Classes/UIKit/手势:触控/手势/MD_GestureRecognizer_VC.m
@@ -104,7 +104,7 @@
 
 -(void)pan:(UIPanGestureRecognizer*)panGR
 {
-    NSLog(@"pan state:%d",panGR.state);//1-began识别成功 2-changed变化中 3-ended撤销
+    NSLog(@"pan state:%ld",(long)panGR.state);//1-began识别成功 2-changed变化中 3-ended撤销
     
     //根据状态改变大小
     if (panGR.state == UIGestureRecognizerStateBegan) {
